myproject/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.0.1
camelot-py==0.9.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
configobj==5.0.9
configparser==7.1.0
cryptography==43.0.1
cymem==2.0.8
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
etelemetry==0.3.1
filelock==3.16.1
fitz==0.0.1.dev2
fsspec==2024.9.0
httplib2==0.22.0
idna==3.10
isodate==0.6.1
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
looseversion==1.3.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.0
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
preshed==3.0.9
prov==2.0.1
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
Pygments==2.18.0
PyMuPDF==1.24.11
pyparsing==3.1.4
PyPDF2==2.10.7
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
requests==2.32.3
rich==13.9.2
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
sympy==1.13.3
thinc==8.3.2
tifffile==2024.9.20
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traits==6.3.2
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
xx_ent_wiki_sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.8.0/xx_ent_wiki_sm-3.8.0-py3-none-any.whl#sha256=6f3c4b853852ea9e9d2dc76cc950dddb10a7e4c42d813308caefe6c5e8be2f0a
